backend/requirements.txt
# Idea Verify Backend Requirements
# Python 3.10+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Data Collection & Scraping
praw==7.7.1              # Reddit API
pygithub==2.1.1          # GitHub API
tweepy==4.14.0           # Twitter API
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Report Generation
reportlab==4.0.8

# Data Analysis & NLP
textblob==0.17.1
vaderSentiment==3.3.2
langdetect==1.0.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# API Documentation
pydantic==2.5.2
pydantic-settings==2.1.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Utilities
httpx==0.25.2
python-dateutil==2.8.2
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
